Move unit tests in line with src code
