+++ src/app/api/v1/__tests__/transactions.test.ts
@@ -1,6 +1,6 @@
 import request from 'supertest'
-import app, { transactionPool } from '../../../app'
-import ValidationTransaction from '../../../transactions/validation'
+import app, { transactionPool } from '../../..'
+import ValidationTransaction from '../../../../transactions/validation'
 
 describe('GET /v1/transactions', () => {
